Enable showing hints for opponent piece moves (#36)
Because hexchess is relatively new, it's hard for players to know where
their opponents' pieces _might_ move.

With the `show-hints` flag enabled, clicking on an opponent's piece,
even if it's not your turn, will display where that piece could move.

Note that it does NOT calculate potentially illegal moves, like moving a
pinned piece, as it is technically not the opponent's turn. It merely
shows all possible moves the piece could potentially do based on its
movement pattern.